Do not debug the compiler build directory on error when --debug is enabled

+++ src/client.ml
@@ -371,11 +371,10 @@ let init_ocaml alias ocaml_version =
       if not (Filename.exists comp) then
         File.Comp.write comp (File.Comp.create_preinstalled ocaml_version)
   with e -> 
-    begin
+    if not !Globals.debug then
       Dirname.rmdir (Path.C.root alias_p);
-      File.Aliases.write aliases_f aliases;
-      raise e;
-    end
+    File.Aliases.write aliases_f aliases;
+    raise e
 
 let indent_left s nb =
   let nb = nb - String.length s in
